move fetch, retrieveAll, query to separate module
